Refactored JobState's meat out into a separate WorkerPool class that might be wrapped in other things like GameSystems, etc..
